working and able to recieve data from database

+++ db/migrations/20180531211016_receipts_table.js
@@ -3,7 +3,7 @@ exports.up = function(knex, Promise) {
     return knex.schema.createTable('receipts', function(t) {
         t.increments('id').primary();
         t.string('location');
-        t.decimal('total');
+        t.integer('total');
         t.date('date');
         t.string('image_url');
     });
